+++ CHANGES.md
@@ -1,3 +1,7 @@
+#### 1.1 / September 26, 2019
+* Made sure shortcode isn't returned on archive pages (see notes in plugin file)
+* Changed the way extra paragraph-tags are removed inside shortcode-tags
+
 #### 1.0.1 / September 17, 2019
 * Minor cleanup en merg of branches
 
@@ -6,7 +10,7 @@
 * Put CSS in <style>-tags in page
 * Removed unnecessary checks
 * Removed ob_start() and put it all into nice strings
-* Removed extra paragraph tags surrounding shortcode-tags
+* Removed extra paragraph-tags surrounding shortcode-tags
 * Changed description to be clearer on usage
 
 #### 0.4.1 / June 8, 2019